Backend/pyproject.toml
[project]
name = "nova-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.3.2",
    "argon2-cffi>=25.1.0",
    "certifi>=2026.1.4",
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "langchain>=1.2.10",
    "langchain-groq>=1.1.2",
    "langchain-mcp-adapters>=0.2.1",
    "langgraph>=1.0.9",
    "passlib>=1.7.4",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.46",
    "sse-starlette>=2.1.0",
    "uuid6>=2025.0.1",
    "uvicorn>=0.41.0",
]
